httpbakery: use an arbitrary number of retries when discharging
Some servers may have a multi-level security system where
the outer level requires a discharge in order to authenticate
and an inner level requires a third party discharge in order
to access a particular thing.

The code now allows an arbitrary number of retries, although 3 is probably
more than sufficient.