httpbakery: make ResponseWriter available in discharger
The identity manager wants to be able to set cookies in
the discharge response, but the current Discharger implementation
does not easily allow third party caveat checkers to do that.

This PR adds a new interface type, ThirdPartyCaveatCheckerP,
that is used in preference to the old checker type, and makes
the response writer (and potentially more things in the future) available.

Technically this change breaks backward compatibility in the bakerytest package,
but our estimation is that this will not break any packages in the wild.